requirements.txt
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
boto3>=1.41.0
strands-agents[anthropic]
strands-agents[gemini]
strands-agents[openai]
strands-agents[a2a]
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
